+++ tools/content-generator/src/intTest/java/org/projectnessie/tools/contentgenerator/ITReadContent.java
@@ -16,9 +16,12 @@
 package org.projectnessie.tools.contentgenerator;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.projectnessie.tools.contentgenerator.ITReadCommits.assertOutputContains;
+import static org.projectnessie.tools.contentgenerator.ITReadCommits.assertOutputDoesNotContain;
 import static org.projectnessie.tools.contentgenerator.RunContentGenerator.runGeneratorCmd;
 
 import java.util.List;
+import java.util.Objects;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -59,7 +62,18 @@ class ITReadContent extends AbstractContentGeneratorTest {
     assertThat(proc).extracting(ProcessResult::getExitCode).isEqualTo(0);
     List<String> output = proc.getStdOutLines();
 
-    assertThat(output).anySatisfy(s -> assertThat(s).contains(CONTENT_KEY.toString()));
+    assertOutputContains(
+        output,
+        "Reading content for key '"
+            + CONTENT_KEY
+            + "' on reference '"
+            + branch.getName()
+            + "' @ HEAD...",
+        "Done reading content for key '"
+            + CONTENT_KEY
+            + "' on reference '"
+            + branch.getName()
+            + "' @ HEAD.");
   }
 
   @Test
@@ -79,31 +93,41 @@ class ITReadContent extends AbstractContentGeneratorTest {
     assertThat(proc).extracting(ProcessResult::getExitCode).isEqualTo(0);
     List<String> output = proc.getStdOutLines();
 
-    assertThat(output).anySatisfy(s -> assertThat(s).contains(CONTENT_KEY.toString()));
-    assertThat(output)
-        .anySatisfy(s -> assertThat(s).contains("key[0]: " + CONTENT_KEY.getElements().get(0)));
-    assertThat(output)
-        .anySatisfy(s -> assertThat(s).contains("key[1]: " + CONTENT_KEY.getElements().get(1)));
+    assertOutputContains(
+        output,
+        "Reading content for key '"
+            + CONTENT_KEY
+            + "' on reference '"
+            + branch.getName()
+            + "' @ HEAD...",
+        "key[0]: " + CONTENT_KEY.getElements().get(0),
+        "key[1]: " + CONTENT_KEY.getElements().get(1),
+        "Done reading content for key '"
+            + CONTENT_KEY
+            + "' on reference '"
+            + branch.getName()
+            + "' @ HEAD.");
   }
 
   @ParameterizedTest
   @CsvSource(
       value = {"%1$s|true", "%2$s|false", "%2$s~1|true"},
       delimiter = '|')
-  void readContentsWithHash(String hash, boolean expectContent) throws Exception {
+  void readContentsWithHash(String hashTemplate, boolean expectContent) throws Exception {
 
     String c1 = branch.getHash();
     try (NessieApiV2 api = buildNessieApi()) {
       branch =
           api.commitMultipleOperations()
               .branchName(branch.getName())
-              .hash(c1)
+              .hash(Objects.requireNonNull(c1))
               .commitMeta(CommitMeta.fromMessage("Second commit"))
               .operation(Operation.Delete.of(CONTENT_KEY))
               .commit();
     }
     String c2 = branch.getHash();
 
+    String hash = String.format(hashTemplate, c1, c2);
     ProcessResult proc =
         runGeneratorCmd(
             "content",
@@ -112,21 +136,33 @@ class ITReadContent extends AbstractContentGeneratorTest {
             "--ref",
             branch.getName(),
             "--hash",
-            String.format(hash, c1, c2),
+            hash,
             "--key",
             CONTENT_KEY.getElements().get(0),
             "--key",
             CONTENT_KEY.getElements().get(1));
     assertThat(proc).extracting(ProcessResult::getExitCode).isEqualTo(0);
     List<String> output = proc.getStdOutLines();
+
+    assertOutputContains(
+        output,
+        "Reading content for key '"
+            + CONTENT_KEY
+            + "' on reference '"
+            + branch.getName()
+            + "' @ "
+            + hash,
+        "Done reading content for key '"
+            + CONTENT_KEY
+            + "' on reference '"
+            + branch.getName()
+            + "' @ "
+            + hash);
+
     if (expectContent) {
-      assertThat(output)
-          .contains("Key: " + CONTENT_KEY)
-          .anySatisfy(s -> assertThat(s).contains("Value: IcebergTable"));
+      assertOutputContains(output, "Key: " + CONTENT_KEY, "Value: IcebergTable");
     } else {
-      assertThat(output)
-          .doesNotContain("Key: " + CONTENT_KEY)
-          .noneSatisfy(s -> assertThat(s).contains("Value: IcebergTable"));
+      assertOutputDoesNotContain(output, "Key: " + CONTENT_KEY, "Value: IcebergTable");
     }
   }
 }